feat(vrang): Moved tests and store models to the correct place Signed-off-by: Stefan Genov <stefan.genov@broadcom.com>
