+++ common/artifact-manager/src/test/java/com/vmware/pscoe/iac/artifact/aria/automation/store/VraNgDay2ActionsPolicyStoreTest.java
@@ -21,6 +21,7 @@ import com.vmware.pscoe.iac.artifact.helpers.FsMocks;
 import com.vmware.pscoe.iac.artifact.model.Package;
 import com.vmware.pscoe.iac.artifact.model.PackageFactory;
 import com.vmware.pscoe.iac.artifact.model.PackageType;
+import com.vmware.pscoe.iac.artifact.aria.automation.store.models.*;
 import com.vmware.pscoe.iac.artifact.aria.automation.models.*;
 import com.vmware.pscoe.iac.artifact.aria.automation.rest.RestClientVraNg;
 import org.junit.Rule;
@@ -44,7 +45,7 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
 import static org.mockito.Mockito.never;
 
-public class VraNgLeasePolicyStoreTest {
+public class VraNgDay2ActionsPolicyStoreTest {
 	/**
 	 * Temp Folder.
 	 */
@@ -53,7 +54,7 @@ public class VraNgLeasePolicyStoreTest {
 	/**
 	 * store.
 	 */
-	private VraNgLeasePolicyStore store;
+	private VraNgDay2ActionsPolicyStore store;
 	/**
 	 * restClient.
 	 */
@@ -77,11 +78,11 @@ public class VraNgLeasePolicyStoreTest {
 	/**
 	 * dirPolicies.
 	 */
-	private final String dirPolicies = "policies";
+	private String dirPolicies = "policies";
 	/**
-	 * Lease Policy.
+	 * day2ActionsPolicy.
 	 */
-	private final String leasePolicy = "lease";
+	private String day2ActionsPolicy = "day2-actions";
 
 	/**
 	 * Init function called before each test.
@@ -95,7 +96,7 @@ public class VraNgLeasePolicyStoreTest {
 		}
 
 		fsMocks = new FsMocks(tempFolder.getRoot());
-		store = new VraNgLeasePolicyStore();
+		store = new VraNgDay2ActionsPolicyStore();
 		restClient = Mockito.mock(RestClientVraNg.class);
 		pkg = PackageFactory.getInstance(PackageType.VRANG, tempFolder.getRoot());
 		config = Mockito.mock(ConfigurationVraNg.class);
@@ -127,31 +128,31 @@ public class VraNgLeasePolicyStoreTest {
 	}
 
 	@Test
-	void testExportContentWithNoLeasePolicies() {
-		System.out.println(this.getClass() + "testExportContentWithNoLeasePolicies");
+	void testExportContentWithNoDay2ActionsPolicies() {
+		System.out.println(this.getClass() + "testExportContentWithNoDay2ActionsPolicies");
 		// GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(new VraNgPolicy());
 
 		// TEST
 		store.exportContent();
 
-		File leasePolicyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, leasePolicy).toFile();
+		File day2ActionsPolicyFolder = Paths
+				.get(tempFolder.getRoot().getPath(), dirPolicies, day2ActionsPolicy).toFile();
 
 		// VERIFY
-		verify(restClient, never()).getLeasePolicies();
-		verify(restClient, never()).getLeasePolicy(anyString());
+		verify(restClient, never()).getDay2ActionsPolicies();
+		verify(restClient, never()).getDay2ActionsPolicy(anyString());
 
-		assertEquals(null, leasePolicyFolder.listFiles());
+		assertEquals(null, day2ActionsPolicyFolder.listFiles());
 	}
 
 	@Test
-	void testExportContentWithAllLeasePolicies() {
-		System.out.println(this.getClass() + "testExportContentWithAllLeasePolicies");
-		VraNgLeasePolicy policy1 = new VraNgLeasePolicy(
+	void testExportContentWithAllDay2ActionsPolicies() {
+		System.out.println(this.getClass() + "testExportContentWithAllDay2ActionsPolicies");
+		VraNgDay2ActionsPolicy policy1 = new VraNgDay2ActionsPolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"LP01",
-				"com.vmware.policy.deployment.lease",
+				"D2A01",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"b2c558c8-f20c-4da6-9bc3-d7561f64df16",
 				"HARD",
@@ -160,10 +161,10 @@ public class VraNgLeasePolicyStoreTest {
 				new JsonObject(),
 				new JsonObject());
 
-		VraNgLeasePolicy policy2 = new VraNgLeasePolicy(
+		VraNgDay2ActionsPolicy policy2 = new VraNgDay2ActionsPolicy(
 				"2cf93725-38e9-4cb9-888a-a40994754c31",
-				"LP02",
-				"com.vmware.policy.deployment.lease",
+				"D2A02",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"b2c558c8-f20c-4da6-9bc3-d7561f64df16",
 				"HARD",
@@ -172,31 +173,31 @@ public class VraNgLeasePolicyStoreTest {
 				new JsonObject(),
 				new JsonObject());
 
-		List<VraNgLeasePolicy> policies = Arrays.asList(policy1, policy2);
+		List<VraNgDay2ActionsPolicy> policies = Arrays.asList(policy1, policy2);
 
 		// // GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(new VraNgPolicy(null, null, null, null, null, null));
-		when(restClient.getLeasePolicies()).thenReturn(policies);
-		when(restClient.getLeasePolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy1);
-		when(restClient.getLeasePolicy("2cf93725-38e9-4cb9-888a-a40994754c31")).thenReturn(policy2);
+		when(restClient.getDay2ActionsPolicies()).thenReturn(policies);
+		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy1);
+		when(restClient.getDay2ActionsPolicy("2cf93725-38e9-4cb9-888a-a40994754c31")).thenReturn(policy2);
 
 		// TEST
 		store.exportContent();
 
 		// VERIFY
-		File policyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, leasePolicy).toFile();
-		assertEquals(2, Objects.requireNonNull(policyFolder.listFiles()).length);
+		File day2ActionsPolicyFolder = Paths
+				.get(tempFolder.getRoot().getPath(), dirPolicies, day2ActionsPolicy).toFile();
+		assertEquals(2, Objects.requireNonNull(day2ActionsPolicyFolder.listFiles()).length);
 	}
 
 	@Test
-	void testExportContentWithSpecificLeasePolicies() {
-		System.out.println(this.getClass() + "testExportContentWithSpecificLeasePolicies");
+	void testExportContentWithSpecificDay2ActionsPolicies() {
+		System.out.println(this.getClass() + "testExportContentWithSpecificDay2ActionsPolicies");
 
-		VraNgLeasePolicy policy = new VraNgLeasePolicy(
+		VraNgDay2ActionsPolicy policy = new VraNgDay2ActionsPolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"LP01",
-				"com.vmware.policy.deployment.lease",
+				"D2A01",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
@@ -204,30 +205,29 @@ public class VraNgLeasePolicyStoreTest {
 				new JsonObject(),
 				new JsonObject(),
 				new JsonObject());
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, null, Collections.singletonList("LP01"), null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, Collections.singletonList("D2A01"), null, null, null);
 		// // GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
-		when(restClient.getLeasePolicies()).thenReturn(Collections.singletonList(policy));
-		when(restClient.getLeasePolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
+		when(restClient.getDay2ActionsPolicies()).thenReturn(Collections.singletonList(policy));
+		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
 
 		// TEST
 		store.exportContent();
 
-		File policyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, leasePolicy).toFile();
+		File day2ActionsPolicyFolder = Paths
+				.get(tempFolder.getRoot().getPath(), dirPolicies, day2ActionsPolicy).toFile();
 
 		// VERIFY
-		assertEquals(1, Objects.requireNonNull(policyFolder.listFiles()).length);
+		assertEquals(1, Objects.requireNonNull(day2ActionsPolicyFolder.listFiles()).length);
 	}
 
 	@Test
 	void testImportContentWithUpdateLogic() {
 		System.out.println("testImportContentWithUpdateLogic");
-		VraNgLeasePolicy policy = new VraNgLeasePolicy(
+		VraNgDay2ActionsPolicy policy = new VraNgDay2ActionsPolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-
-				"LP01",
-				"com.vmware.policy.deployment.lease",
+				"D2A01",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
@@ -235,36 +235,36 @@ public class VraNgLeasePolicyStoreTest {
 				new JsonObject(),
 				new JsonObject(),
 				new JsonObject());
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, null, Collections.singletonList("LP01"), null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, Collections.singletonList("D2A01"), null, null, null);
 		// GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
 
-		fsMocks.getLeasePolicyFsMocks().addPolicy(policy);
+		fsMocks.getDay2ActionsPolicyFsMocks().addPolicy(policy);
 
-		File policyFolder = Paths
-				.get(fsMocks.getTempFolderProjectPath().getPath(), leasePolicy).toFile();
+		File day2ActionsPolicyFolder = Paths
+				.get(fsMocks.getTempFolderProjectPath().getPath(), day2ActionsPolicy).toFile();
 
-		AssertionsHelper.assertFolderContainsFiles(policyFolder, new String[] { "LP01.json" });
+		AssertionsHelper.assertFolderContainsFiles(day2ActionsPolicyFolder, new String[] { "D2A01.json" });
 
-		when(restClient.getLeasePolicies()).thenReturn(Arrays.asList(policy));
-		when(restClient.getLeasePolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
+		when(restClient.getDay2ActionsPolicies()).thenReturn(Arrays.asList(policy));
+		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
 
 		// START TEST
 		store.importContent(tempFolder.getRoot());
 
 		// VERIFY
-		verify(restClient, times(1)).createLeasePolicy(any());
+		verify(restClient, times(1)).createDay2ActionsPolicy(any());
 	}
 
 	@Test
 	void testImportContentWithCreateLogic() {
 		System.out.println("testImportContentWithCreateLogic");
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, null, Collections.singletonList("LP01"), null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, Arrays.asList("D2A01"), null, null, null);
 
-		VraNgLeasePolicy policy = new VraNgLeasePolicy(
+		VraNgDay2ActionsPolicy policy = new VraNgDay2ActionsPolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"LP01",
-				"com.vmware.policy.deployment.lease",
+				"D2A01",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"b2c558c8-f20c-4da6-9bc3-d7561f64df16",
 				"HARD",
@@ -272,10 +272,10 @@ public class VraNgLeasePolicyStoreTest {
 				new JsonObject(),
 				new JsonObject(),
 				new JsonObject());
-		VraNgLeasePolicy policyFromServer = new VraNgLeasePolicy(
+		VraNgDay2ActionsPolicy policyFromServer = new VraNgDay2ActionsPolicy(
 				"2cf93725-38e9-4cb9-888a-a40994754c31",
-				"LP02",
-				"com.vmware.policy.deployment.lease",
+				"D2A02",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"b2c558c8-f20c-4da6-9bc3-d7561f64df16",
 				"HARD",
@@ -287,21 +287,21 @@ public class VraNgLeasePolicyStoreTest {
 		// GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
 
-		fsMocks.getLeasePolicyFsMocks().addPolicy(policy);
+		fsMocks.getDay2ActionsPolicyFsMocks().addPolicy(policy);
 
-		File policyFolder = Paths
-				.get(fsMocks.getTempFolderProjectPath().getPath(), leasePolicy).toFile();
+		File day2ActionsPolicyFolder = Paths
+				.get(fsMocks.getTempFolderProjectPath().getPath(), day2ActionsPolicy).toFile();
 
-		AssertionsHelper.assertFolderContainsFiles(policyFolder, new String[] { "LP01.json" });
+		AssertionsHelper.assertFolderContainsFiles(day2ActionsPolicyFolder, new String[] { "D2A01.json" });
 
-		when(restClient.getLeasePolicies()).thenReturn(Collections.emptyList());
-		when(restClient.getLeasePolicy("2cf93725-38e9-4cb9-888a-a40994754c31")).thenReturn(policyFromServer);
+		when(restClient.getDay2ActionsPolicies()).thenReturn(Collections.emptyList());
+		when(restClient.getDay2ActionsPolicy("2cf93725-38e9-4cb9-888a-a40994754c31")).thenReturn(policyFromServer);
 
 		// START TEST
 		store.importContent(tempFolder.getRoot());
 
 		// VERIFY
-		verify(restClient, times(1)).createLeasePolicy(any());
+		verify(restClient, times(1)).createDay2ActionsPolicy(any());
 	}
 
 	@Test
@@ -313,19 +313,19 @@ public class VraNgLeasePolicyStoreTest {
 
 		// VERIFY
 		verify(vraNgPackageDescriptor, never()).getPolicy();
-		verify(restClient, never()).getLeasePolicies();
-		verify(restClient, never()).getLeasePolicy(anyString());
-		verify(restClient, never()).createLeasePolicy(any());
+		verify(restClient, never()).getDay2ActionsPolicies();
+		verify(restClient, never()).getDay2ActionsPolicy(anyString());
+		verify(restClient, never()).createDay2ActionsPolicy(any());
 	}
 
 	@Test
 	void testExportContentWithPolicyAlreadyInFile() {
 		System.out.println(this.getClass() + ".testExportContentWithPolicyAlreadyInFile");
 
-		VraNgLeasePolicy policy = new VraNgLeasePolicy(
+		VraNgDay2ActionsPolicy policy = new VraNgDay2ActionsPolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"LP01",
-				"com.vmware.policy.deployment.lease",
+				"D2A01",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
@@ -333,16 +333,16 @@ public class VraNgLeasePolicyStoreTest {
 				new JsonObject(),
 				new JsonObject(),
 				new JsonObject());
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, null, Collections.singletonList("LP01"), null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, Collections.singletonList("D2A01"), null, null, null);
 		// // GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
-		when(restClient.getLeasePolicies()).thenReturn(Collections.singletonList(policy));
-		when(restClient.getLeasePolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
+		when(restClient.getDay2ActionsPolicies()).thenReturn(Collections.singletonList(policy));
+		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
 
 		File policyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, leasePolicy).toFile();
+				.get(tempFolder.getRoot().getPath(), dirPolicies, day2ActionsPolicy).toFile();
 
-		fsMocks.getLeasePolicyFsMocks().addPolicy(policy);
+		fsMocks.getDay2ActionsPolicyFsMocks().addPolicy(policy);
 		// TEST
 		store.exportContent();
 
@@ -354,10 +354,10 @@ public class VraNgLeasePolicyStoreTest {
 	@Test
 	void testExportContentWithSpecificPoliciesAndDuplicateFiles() {
 		System.out.println(this.getClass() + ".testExportContentWithSpecificPoliciesAndDuplicateFiles");
-		VraNgLeasePolicy policyInFile = new VraNgLeasePolicy(
+		VraNgDay2ActionsPolicy policyInFile = new VraNgDay2ActionsPolicy(
 				"d160119e-4027-48d1-a2b5-5229b3cee282",
-				"LP01",
-				"com.vmware.policy.deployment.lease",
+				"D2A01",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"SOFT",
@@ -366,10 +366,10 @@ public class VraNgLeasePolicyStoreTest {
 				new JsonObject(),
 				new JsonObject());
 
-		VraNgLeasePolicy policy = new VraNgLeasePolicy(
+		VraNgDay2ActionsPolicy policy = new VraNgDay2ActionsPolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"LP01",
-				"com.vmware.policy.deployment.lease",
+				"D2A01",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
@@ -377,22 +377,22 @@ public class VraNgLeasePolicyStoreTest {
 				new JsonObject(),
 				new JsonObject(),
 				new JsonObject());
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, null, Collections.singletonList("LP01"), null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, Collections.singletonList("D2A01"), null, null, null);
 		// // GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
-		when(restClient.getLeasePolicies()).thenReturn(Collections.singletonList(policy));
-		when(restClient.getLeasePolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
+		when(restClient.getDay2ActionsPolicies()).thenReturn(Collections.singletonList(policy));
+		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
 
 		File policyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, leasePolicy).toFile();
+				.get(tempFolder.getRoot().getPath(), dirPolicies, day2ActionsPolicy).toFile();
 
-		fsMocks.getLeasePolicyFsMocks().addPolicy(policyInFile);
-		policyInFile.setName("LP01_1");
-		fsMocks.getLeasePolicyFsMocks().addPolicy(policyInFile);
-		policyInFile.setName("LP01_2");
-		fsMocks.getLeasePolicyFsMocks().addPolicy(policyInFile);
-		policyInFile.setName("LP01_3");
-		fsMocks.getLeasePolicyFsMocks().addPolicy(policyInFile);
+		fsMocks.getDay2ActionsPolicyFsMocks().addPolicy(policyInFile);
+		policyInFile.setName("D2A01_1");
+		fsMocks.getDay2ActionsPolicyFsMocks().addPolicy(policyInFile);
+		policyInFile.setName("D2A01_2");
+		fsMocks.getDay2ActionsPolicyFsMocks().addPolicy(policyInFile);
+		policyInFile.setName("D2A01_3");
+		fsMocks.getDay2ActionsPolicyFsMocks().addPolicy(policyInFile);
 
 		// TEST
 		store.exportContent();
@@ -400,16 +400,16 @@ public class VraNgLeasePolicyStoreTest {
 		// VERIFY
 		assertEquals(5, Objects.requireNonNull(policyFolder.listFiles()).length);
 		AssertionsHelper.assertFolderContainsFiles(policyFolder,
-				new String[] { "LP01.json", "LP01_1.json", "LP01_2.json", "LP01_3.json", "LP01_4.json" });
+				new String[] { "D2A01.json", "D2A01_1.json", "D2A01_2.json", "D2A01_3.json", "D2A01_4.json" });
 	}
 
 	@Test
 	void testExportContentWithSpecificPoliciesAndDuplicateNames() {
 		System.out.println(this.getClass() + ".testExportContentWithSpecificPoliciesAndDuplicateNames");
-		VraNgLeasePolicy policyInFile = new VraNgLeasePolicy(
+		VraNgDay2ActionsPolicy policyInFile = new VraNgDay2ActionsPolicy(
 				"d160119e-4027-48d1-a2b5-5229b3cee282",
-				"LP01",
-				"com.vmware.policy.deployment.lease",
+				"D2A01",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"SOFT",
@@ -418,10 +418,10 @@ public class VraNgLeasePolicyStoreTest {
 				new JsonObject(),
 				new JsonObject());
 
-		VraNgLeasePolicy policy = new VraNgLeasePolicy(
+		VraNgDay2ActionsPolicy policy = new VraNgDay2ActionsPolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"LP01",
-				"com.vmware.policy.deployment.lease",
+				"D2A01",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
@@ -429,10 +429,10 @@ public class VraNgLeasePolicyStoreTest {
 				new JsonObject(),
 				new JsonObject(),
 				new JsonObject());
-		VraNgLeasePolicy policy1 = new VraNgLeasePolicy(
+		VraNgDay2ActionsPolicy policy1 = new VraNgDay2ActionsPolicy(
 				"df60ff9e-4027-11d1-a2b5-5229b3cee282",
-				"LP01",
-				"com.vmware.policy.deployment.lease",
+				"D2A01",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
@@ -440,10 +440,10 @@ public class VraNgLeasePolicyStoreTest {
 				new JsonObject(),
 				new JsonObject(),
 				new JsonObject());
-		VraNgLeasePolicy policy2 = new VraNgLeasePolicy(
+		VraNgDay2ActionsPolicy policy2 = new VraNgDay2ActionsPolicy(
 				"df60ff9e-4027-12d1-a2b5-5229b3cee282",
-				"LP01",
-				"com.vmware.policy.deployment.lease",
+				"D2A01",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
@@ -451,10 +451,10 @@ public class VraNgLeasePolicyStoreTest {
 				new JsonObject(),
 				new JsonObject(),
 				new JsonObject());
-		VraNgLeasePolicy policy3 = new VraNgLeasePolicy(
+		VraNgDay2ActionsPolicy policy3 = new VraNgDay2ActionsPolicy(
 				"df60ff9e-4027-13d1-a2b5-5229b3cee282",
-				"LP01",
-				"com.vmware.policy.deployment.lease",
+				"D2A01",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
@@ -462,10 +462,10 @@ public class VraNgLeasePolicyStoreTest {
 				new JsonObject(),
 				new JsonObject(),
 				new JsonObject());
-		VraNgLeasePolicy policy4 = new VraNgLeasePolicy(
+		VraNgDay2ActionsPolicy policy4 = new VraNgDay2ActionsPolicy(
 				"df60ff9e-4027-14d1-a2b5-5229b3cee282",
-				"LP01",
-				"com.vmware.policy.deployment.lease",
+				"D2A01",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
@@ -473,10 +473,10 @@ public class VraNgLeasePolicyStoreTest {
 				new JsonObject(),
 				new JsonObject(),
 				new JsonObject());
-		VraNgLeasePolicy policy5 = new VraNgLeasePolicy(
+		VraNgDay2ActionsPolicy policy5 = new VraNgDay2ActionsPolicy(
 				"df60ff9e-4027-15d1-a2b5-5229b3cee282",
-				"LP01",
-				"com.vmware.policy.deployment.lease",
+				"D2A01",
+				"com.vmware.policy.deployment.action",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
@@ -485,27 +485,28 @@ public class VraNgLeasePolicyStoreTest {
 				new JsonObject(),
 				new JsonObject());
 
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, null, Collections.singletonList("LP01"), null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, Arrays.asList("D2A01"), null, null, null);
 		// // GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
-		when(restClient.getLeasePolicies())
+		when(restClient.getDay2ActionsPolicies())
 				.thenReturn(Arrays.asList(policy, policy1, policy2, policy3, policy4, policy5));
-		when(restClient.getLeasePolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
-		when(restClient.getLeasePolicy("df60ff9e-4027-11d1-a2b5-5229b3cee282")).thenReturn(policy1);
-		when(restClient.getLeasePolicy("df60ff9e-4027-12d1-a2b5-5229b3cee282")).thenReturn(policy2);
-		when(restClient.getLeasePolicy("df60ff9e-4027-13d1-a2b5-5229b3cee282")).thenReturn(policy3);
-		when(restClient.getLeasePolicy("df60ff9e-4027-14d1-a2b5-5229b3cee282")).thenReturn(policy4);
-		when(restClient.getLeasePolicy("df60ff9e-4027-15d1-a2b5-5229b3cee282")).thenReturn(policy5);
+		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
+		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-11d1-a2b5-5229b3cee282")).thenReturn(policy1);
+		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-12d1-a2b5-5229b3cee282")).thenReturn(policy2);
+		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-13d1-a2b5-5229b3cee282")).thenReturn(policy3);
+		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-14d1-a2b5-5229b3cee282")).thenReturn(policy4);
+		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-15d1-a2b5-5229b3cee282")).thenReturn(policy5);
 
 		File policyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, leasePolicy).toFile();
+				.get(tempFolder.getRoot().getPath(), dirPolicies, day2ActionsPolicy).toFile();
 
 		// TEST
 		store.exportContent();
 
 		// VERIFY
 		assertEquals(6, Objects.requireNonNull(policyFolder.listFiles()).length);
-		AssertionsHelper.assertFolderContainsFiles(policyFolder, new String[] { "LP01.json", "LP01_1.json",
-				"LP01_2.json", "LP01_3.json", "LP01_4.json", "LP01_5.json" });
+		AssertionsHelper.assertFolderContainsFiles(policyFolder, new String[] { "D2A01.json", "D2A01_1.json",
+				"D2A01_2.json", "D2A01_3.json", "D2A01_4.json", "D2A01_5.json" });
 	}
+
 }
