+++ common/artifact-manager/src/test/java/com/vmware/pscoe/iac/artifact/aria/automation/store/VraNgLeasePolicyStoreTest.java
@@ -21,6 +21,7 @@ import com.vmware.pscoe.iac.artifact.helpers.FsMocks;
 import com.vmware.pscoe.iac.artifact.model.Package;
 import com.vmware.pscoe.iac.artifact.model.PackageFactory;
 import com.vmware.pscoe.iac.artifact.model.PackageType;
+import com.vmware.pscoe.iac.artifact.aria.automation.store.models.*;
 import com.vmware.pscoe.iac.artifact.aria.automation.models.*;
 import com.vmware.pscoe.iac.artifact.aria.automation.rest.RestClientVraNg;
 import org.junit.Rule;
@@ -33,15 +34,18 @@ import org.mockito.Mockito;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Paths;
-import java.util.*;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.Objects;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.never;
 
-public class VraNgResourceQuotaPolicyStoreTest {
-
+public class VraNgLeasePolicyStoreTest {
 	/**
 	 * Temp Folder.
 	 */
@@ -50,7 +54,7 @@ public class VraNgResourceQuotaPolicyStoreTest {
 	/**
 	 * store.
 	 */
-	private VraNgResourceQuotaPolicyStore store;
+	private VraNgLeasePolicyStore store;
 	/**
 	 * restClient.
 	 */
@@ -74,11 +78,11 @@ public class VraNgResourceQuotaPolicyStoreTest {
 	/**
 	 * dirPolicies.
 	 */
-	private String dirPolicies = "policies";
+	private final String dirPolicies = "policies";
 	/**
-	 * resourceQuotaPolicy.
+	 * Lease Policy.
 	 */
-	private String resourceQuotaPolicy = "resource-quota";
+	private final String leasePolicy = "lease";
 
 	/**
 	 * Init function called before each test.
@@ -92,7 +96,7 @@ public class VraNgResourceQuotaPolicyStoreTest {
 		}
 
 		fsMocks = new FsMocks(tempFolder.getRoot());
-		store = new VraNgResourceQuotaPolicyStore();
+		store = new VraNgLeasePolicyStore();
 		restClient = Mockito.mock(RestClientVraNg.class);
 		pkg = PackageFactory.getInstance(PackageType.VRANG, tempFolder.getRoot());
 		config = Mockito.mock(ConfigurationVraNg.class);
@@ -124,174 +128,181 @@ public class VraNgResourceQuotaPolicyStoreTest {
 	}
 
 	@Test
-	void testExportContentWithNoResourceQuotaPolicies() {
-		System.out.println(this.getClass() + "testExportContentWithNoResourceQuotaPolicies");
+	void testExportContentWithNoLeasePolicies() {
+		System.out.println(this.getClass() + "testExportContentWithNoLeasePolicies");
 		// GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(new VraNgPolicy());
 
 		// TEST
 		store.exportContent();
 
-		File resourceQuotaPolicyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, resourceQuotaPolicy).toFile();
+		File leasePolicyFolder = Paths
+				.get(tempFolder.getRoot().getPath(), dirPolicies, leasePolicy).toFile();
 
 		// VERIFY
-		verify(restClient, never()).getResourceQuotaPolicies();
-		verify(restClient, never()).getResourceQuotaPolicy(anyString());
+		verify(restClient, never()).getLeasePolicies();
+		verify(restClient, never()).getLeasePolicy(anyString());
 
-		assertEquals(null, resourceQuotaPolicyFolder.listFiles());
+		assertEquals(null, leasePolicyFolder.listFiles());
 	}
 
 	@Test
-	void testExportContentWithAllResourceQuotaPolicies() {
-		System.out.println(this.getClass() + "testExportContentWithAllResourceQuotaPolicies");
-		VraNgResourceQuotaPolicy rqPolicy = new VraNgResourceQuotaPolicy(
+	void testExportContentWithAllLeasePolicies() {
+		System.out.println(this.getClass() + "testExportContentWithAllLeasePolicies");
+		VraNgLeasePolicy policy1 = new VraNgLeasePolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"RQ01",
-				"com.vmware.policy.resource.quota",
+				"LP01",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"b2c558c8-f20c-4da6-9bc3-d7561f64df16",
 				"HARD",
 				"TEST",
 				new JsonObject(),
+				new JsonObject(),
 				new JsonObject());
 
-		VraNgResourceQuotaPolicy rqPolicy2 = new VraNgResourceQuotaPolicy(
+		VraNgLeasePolicy policy2 = new VraNgLeasePolicy(
 				"2cf93725-38e9-4cb9-888a-a40994754c31",
-				"RQ02",
-				"com.vmware.policy.resource.quota",
+				"LP02",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"b2c558c8-f20c-4da6-9bc3-d7561f64df16",
 				"HARD",
 				"TEST",
 				new JsonObject(),
+				new JsonObject(),
 				new JsonObject());
 
-		List<VraNgResourceQuotaPolicy> policies = Arrays.asList(rqPolicy, rqPolicy2);
+		List<VraNgLeasePolicy> policies = Arrays.asList(policy1, policy2);
 
 		// // GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(new VraNgPolicy(null, null, null, null, null, null));
-		when(restClient.getResourceQuotaPolicies()).thenReturn(policies);
-		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(rqPolicy);
-		when(restClient.getResourceQuotaPolicy("2cf93725-38e9-4cb9-888a-a40994754c31")).thenReturn(rqPolicy2);
+		when(restClient.getLeasePolicies()).thenReturn(policies);
+		when(restClient.getLeasePolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy1);
+		when(restClient.getLeasePolicy("2cf93725-38e9-4cb9-888a-a40994754c31")).thenReturn(policy2);
 
 		// TEST
 		store.exportContent();
 
 		// VERIFY
-		File resourceQuotaPolicyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, resourceQuotaPolicy).toFile();
-		assertEquals(2, Objects.requireNonNull(resourceQuotaPolicyFolder.listFiles()).length);
+		File policyFolder = Paths
+				.get(tempFolder.getRoot().getPath(), dirPolicies, leasePolicy).toFile();
+		assertEquals(2, Objects.requireNonNull(policyFolder.listFiles()).length);
 	}
 
 	@Test
-	void testExportContentWithSpecificResourceQuotaPolicies() {
-		System.out.println(this.getClass() + "testExportContentWithSpecificResourceQuotaPolicies");
+	void testExportContentWithSpecificLeasePolicies() {
+		System.out.println(this.getClass() + "testExportContentWithSpecificLeasePolicies");
 
-		VraNgResourceQuotaPolicy rqPolicy = new VraNgResourceQuotaPolicy(
+		VraNgLeasePolicy policy = new VraNgLeasePolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"RQ01",
-				"com.vmware.policy.resource.quota",
+				"LP01",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST",
 				new JsonObject(),
+				new JsonObject(),
 				new JsonObject());
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, Collections.singletonList("RQ01"), null, null, null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, null, Collections.singletonList("LP01"), null, null);
 		// // GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
-		when(restClient.getResourceQuotaPolicies()).thenReturn(Collections.singletonList(rqPolicy));
-		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(rqPolicy);
+		when(restClient.getLeasePolicies()).thenReturn(Collections.singletonList(policy));
+		when(restClient.getLeasePolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
 
 		// TEST
 		store.exportContent();
 
-		File resourceQuotaPolicyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, resourceQuotaPolicy).toFile();
+		File policyFolder = Paths
+				.get(tempFolder.getRoot().getPath(), dirPolicies, leasePolicy).toFile();
 
 		// VERIFY
-		assertEquals(1, Objects.requireNonNull(resourceQuotaPolicyFolder.listFiles()).length);
+		assertEquals(1, Objects.requireNonNull(policyFolder.listFiles()).length);
 	}
 
 	@Test
 	void testImportContentWithUpdateLogic() {
 		System.out.println("testImportContentWithUpdateLogic");
-		VraNgResourceQuotaPolicy rqPolicy = new VraNgResourceQuotaPolicy(
+		VraNgLeasePolicy policy = new VraNgLeasePolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"RQ01",
-				"com.vmware.policy.resource.quota",
+
+				"LP01",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST",
 				new JsonObject(),
+				new JsonObject(),
 				new JsonObject());
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, Collections.singletonList("RQ01"), null, null, null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, null, Collections.singletonList("LP01"), null, null);
 		// GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
 
-		fsMocks.resourceQuotaPolicyFsMocks().addResourceQuotaPolicy(rqPolicy);
+		fsMocks.getLeasePolicyFsMocks().addPolicy(policy);
 
-		File resourceQuotaPolicyFolder = Paths
-				.get(fsMocks.getTempFolderProjectPath().getPath(), resourceQuotaPolicy).toFile();
+		File policyFolder = Paths
+				.get(fsMocks.getTempFolderProjectPath().getPath(), leasePolicy).toFile();
 
-		AssertionsHelper.assertFolderContainsFiles(resourceQuotaPolicyFolder, new String[] { "RQ01.json" });
+		AssertionsHelper.assertFolderContainsFiles(policyFolder, new String[] { "LP01.json" });
 
-		when(restClient.getResourceQuotaPolicies()).thenReturn(Arrays.asList(rqPolicy));
-		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(rqPolicy);
+		when(restClient.getLeasePolicies()).thenReturn(Arrays.asList(policy));
+		when(restClient.getLeasePolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
 
 		// START TEST
 		store.importContent(tempFolder.getRoot());
 
 		// VERIFY
-		verify(restClient, times(1)).createResourceQuotaPolicy(any());
+		verify(restClient, times(1)).createLeasePolicy(any());
 	}
 
 	@Test
 	void testImportContentWithCreateLogic() {
 		System.out.println("testImportContentWithCreateLogic");
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, Collections.singletonList("RQ01"), null, null, null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, null, Collections.singletonList("LP01"), null, null);
 
-		VraNgResourceQuotaPolicy rqPolicy = new VraNgResourceQuotaPolicy(
+		VraNgLeasePolicy policy = new VraNgLeasePolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"RQ01",
-				"com.vmware.policy.resource.quota",
+				"LP01",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"b2c558c8-f20c-4da6-9bc3-d7561f64df16",
 				"HARD",
 				"TEST",
 				new JsonObject(),
+				new JsonObject(),
 				new JsonObject());
-		VraNgResourceQuotaPolicy rqPolicyFromServer = new VraNgResourceQuotaPolicy(
+		VraNgLeasePolicy policyFromServer = new VraNgLeasePolicy(
 				"2cf93725-38e9-4cb9-888a-a40994754c31",
-				"RQ02",
-				"com.vmware.policy.resource.quota",
+				"LP02",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"b2c558c8-f20c-4da6-9bc3-d7561f64df16",
 				"HARD",
 				"TEST",
+				new JsonObject(),
 				new JsonObject(),
 				new JsonObject());
 
 		// GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
 
-		fsMocks.resourceQuotaPolicyFsMocks().addResourceQuotaPolicy(rqPolicy);
+		fsMocks.getLeasePolicyFsMocks().addPolicy(policy);
 
-		File resourceQuotaPolicyFolder = Paths
-				.get(fsMocks.getTempFolderProjectPath().getPath(), resourceQuotaPolicy).toFile();
+		File policyFolder = Paths
+				.get(fsMocks.getTempFolderProjectPath().getPath(), leasePolicy).toFile();
 
-		AssertionsHelper.assertFolderContainsFiles(resourceQuotaPolicyFolder, new String[] { "RQ01.json" });
+		AssertionsHelper.assertFolderContainsFiles(policyFolder, new String[] { "LP01.json" });
 
-		when(restClient.getResourceQuotaPolicies()).thenReturn(Collections.emptyList());
-		when(restClient.getResourceQuotaPolicy("2cf93725-38e9-4cb9-888a-a40994754c31")).thenReturn(rqPolicyFromServer);
+		when(restClient.getLeasePolicies()).thenReturn(Collections.emptyList());
+		when(restClient.getLeasePolicy("2cf93725-38e9-4cb9-888a-a40994754c31")).thenReturn(policyFromServer);
 
 		// START TEST
 		store.importContent(tempFolder.getRoot());
 
 		// VERIFY
-		verify(restClient, times(1)).createResourceQuotaPolicy(any());
+		verify(restClient, times(1)).createLeasePolicy(any());
 	}
 
 	@Test
@@ -303,35 +314,36 @@ public class VraNgResourceQuotaPolicyStoreTest {
 
 		// VERIFY
 		verify(vraNgPackageDescriptor, never()).getPolicy();
-		verify(restClient, never()).getResourceQuotaPolicies();
-		verify(restClient, never()).getResourceQuotaPolicy(anyString());
-		verify(restClient, never()).createResourceQuotaPolicy(any());
+		verify(restClient, never()).getLeasePolicies();
+		verify(restClient, never()).getLeasePolicy(anyString());
+		verify(restClient, never()).createLeasePolicy(any());
 	}
 
 	@Test
 	void testExportContentWithPolicyAlreadyInFile() {
 		System.out.println(this.getClass() + ".testExportContentWithPolicyAlreadyInFile");
 
-		VraNgResourceQuotaPolicy policy = new VraNgResourceQuotaPolicy(
+		VraNgLeasePolicy policy = new VraNgLeasePolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"RQ01",
-				"com.vmware.policy.resource.quota",
+				"LP01",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST",
 				new JsonObject(),
+				new JsonObject(),
 				new JsonObject());
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, Collections.singletonList("RQ01"), null, null, null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, null, Collections.singletonList("LP01"), null, null);
 		// // GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
-		when(restClient.getResourceQuotaPolicies()).thenReturn(Collections.singletonList(policy));
-		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
+		when(restClient.getLeasePolicies()).thenReturn(Collections.singletonList(policy));
+		when(restClient.getLeasePolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
 
 		File policyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, resourceQuotaPolicy).toFile();
+				.get(tempFolder.getRoot().getPath(), dirPolicies, leasePolicy).toFile();
 
-		fsMocks.resourceQuotaPolicyFsMocks().addResourceQuotaPolicy(policy);
+		fsMocks.getLeasePolicyFsMocks().addPolicy(policy);
 		// TEST
 		store.exportContent();
 
@@ -343,43 +355,45 @@ public class VraNgResourceQuotaPolicyStoreTest {
 	@Test
 	void testExportContentWithSpecificPoliciesAndDuplicateFiles() {
 		System.out.println(this.getClass() + ".testExportContentWithSpecificPoliciesAndDuplicateFiles");
-		VraNgResourceQuotaPolicy policyInFile = new VraNgResourceQuotaPolicy(
+		VraNgLeasePolicy policyInFile = new VraNgLeasePolicy(
 				"d160119e-4027-48d1-a2b5-5229b3cee282",
-				"RQ01",
-				"com.vmware.policy.resource.quota",
+				"LP01",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"SOFT",
 				"TEST",
 				new JsonObject(),
+				new JsonObject(),
 				new JsonObject());
 
-		VraNgResourceQuotaPolicy policy = new VraNgResourceQuotaPolicy(
+		VraNgLeasePolicy policy = new VraNgLeasePolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"RQ01",
-				"com.vmware.policy.resource.quota",
+				"LP01",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST",
 				new JsonObject(),
+				new JsonObject(),
 				new JsonObject());
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, Collections.singletonList("RQ01"), null, null, null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, null, Collections.singletonList("LP01"), null, null);
 		// // GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
-		when(restClient.getResourceQuotaPolicies()).thenReturn(Collections.singletonList(policy));
-		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
+		when(restClient.getLeasePolicies()).thenReturn(Collections.singletonList(policy));
+		when(restClient.getLeasePolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
 
 		File policyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, resourceQuotaPolicy).toFile();
+				.get(tempFolder.getRoot().getPath(), dirPolicies, leasePolicy).toFile();
 
-		fsMocks.resourceQuotaPolicyFsMocks().addResourceQuotaPolicy(policyInFile);
-		policyInFile.setName("RQ01_1");
-		fsMocks.resourceQuotaPolicyFsMocks().addResourceQuotaPolicy(policyInFile);
-		policyInFile.setName("RQ01_2");
-		fsMocks.resourceQuotaPolicyFsMocks().addResourceQuotaPolicy(policyInFile);
-		policyInFile.setName("RQ01_3");
-		fsMocks.resourceQuotaPolicyFsMocks().addResourceQuotaPolicy(policyInFile);
+		fsMocks.getLeasePolicyFsMocks().addPolicy(policyInFile);
+		policyInFile.setName("LP01_1");
+		fsMocks.getLeasePolicyFsMocks().addPolicy(policyInFile);
+		policyInFile.setName("LP01_2");
+		fsMocks.getLeasePolicyFsMocks().addPolicy(policyInFile);
+		policyInFile.setName("LP01_3");
+		fsMocks.getLeasePolicyFsMocks().addPolicy(policyInFile);
 
 		// TEST
 		store.exportContent();
@@ -387,105 +401,112 @@ public class VraNgResourceQuotaPolicyStoreTest {
 		// VERIFY
 		assertEquals(5, Objects.requireNonNull(policyFolder.listFiles()).length);
 		AssertionsHelper.assertFolderContainsFiles(policyFolder,
-				new String[] { "RQ01.json", "RQ01_1.json", "RQ01_2.json", "RQ01_3.json", "RQ01_4.json" });
+				new String[] { "LP01.json", "LP01_1.json", "LP01_2.json", "LP01_3.json", "LP01_4.json" });
 	}
 
 	@Test
 	void testExportContentWithSpecificPoliciesAndDuplicateNames() {
 		System.out.println(this.getClass() + ".testExportContentWithSpecificPoliciesAndDuplicateNames");
-		VraNgResourceQuotaPolicy policyInFile = new VraNgResourceQuotaPolicy(
+		VraNgLeasePolicy policyInFile = new VraNgLeasePolicy(
 				"d160119e-4027-48d1-a2b5-5229b3cee282",
-				"RQ01",
-				"com.vmware.policy.resource.quota",
+				"LP01",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"SOFT",
 				"TEST",
 				new JsonObject(),
+				new JsonObject(),
 				new JsonObject());
 
-		VraNgResourceQuotaPolicy policy = new VraNgResourceQuotaPolicy(
+		VraNgLeasePolicy policy = new VraNgLeasePolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"RQ01",
-				"com.vmware.policy.resource.quota",
+				"LP01",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST",
 				new JsonObject(),
+				new JsonObject(),
 				new JsonObject());
-		VraNgResourceQuotaPolicy policy1 = new VraNgResourceQuotaPolicy(
+		VraNgLeasePolicy policy1 = new VraNgLeasePolicy(
 				"df60ff9e-4027-11d1-a2b5-5229b3cee282",
-				"RQ01",
-				"com.vmware.policy.resource.quota",
+				"LP01",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST1",
 				new JsonObject(),
+				new JsonObject(),
 				new JsonObject());
-		VraNgResourceQuotaPolicy policy2 = new VraNgResourceQuotaPolicy(
+		VraNgLeasePolicy policy2 = new VraNgLeasePolicy(
 				"df60ff9e-4027-12d1-a2b5-5229b3cee282",
-				"RQ01",
-				"com.vmware.policy.resource.quota",
+				"LP01",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST2",
 				new JsonObject(),
+				new JsonObject(),
 				new JsonObject());
-		VraNgResourceQuotaPolicy policy3 = new VraNgResourceQuotaPolicy(
+		VraNgLeasePolicy policy3 = new VraNgLeasePolicy(
 				"df60ff9e-4027-13d1-a2b5-5229b3cee282",
-				"RQ01",
-				"com.vmware.policy.resource.quota",
+				"LP01",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST3",
 				new JsonObject(),
+				new JsonObject(),
 				new JsonObject());
-		VraNgResourceQuotaPolicy policy4 = new VraNgResourceQuotaPolicy(
+		VraNgLeasePolicy policy4 = new VraNgLeasePolicy(
 				"df60ff9e-4027-14d1-a2b5-5229b3cee282",
-				"RQ01",
-				"com.vmware.policy.resource.quota",
+				"LP01",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST4",
 				new JsonObject(),
+				new JsonObject(),
 				new JsonObject());
-		VraNgResourceQuotaPolicy policy5 = new VraNgResourceQuotaPolicy(
+		VraNgLeasePolicy policy5 = new VraNgLeasePolicy(
 				"df60ff9e-4027-15d1-a2b5-5229b3cee282",
-				"RQ01",
-				"com.vmware.policy.resource.quota",
+				"LP01",
+				"com.vmware.policy.deployment.lease",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST5",
 				new JsonObject(),
+				new JsonObject(),
 				new JsonObject());
 
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, Collections.singletonList("RQ01"), null, null, null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, null, Collections.singletonList("LP01"), null, null);
 		// // GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
-		when(restClient.getResourceQuotaPolicies())
+		when(restClient.getLeasePolicies())
 				.thenReturn(Arrays.asList(policy, policy1, policy2, policy3, policy4, policy5));
-		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
-		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-11d1-a2b5-5229b3cee282")).thenReturn(policy1);
-		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-12d1-a2b5-5229b3cee282")).thenReturn(policy2);
-		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-13d1-a2b5-5229b3cee282")).thenReturn(policy3);
-		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-14d1-a2b5-5229b3cee282")).thenReturn(policy4);
-		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-15d1-a2b5-5229b3cee282")).thenReturn(policy5);
+		when(restClient.getLeasePolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
+		when(restClient.getLeasePolicy("df60ff9e-4027-11d1-a2b5-5229b3cee282")).thenReturn(policy1);
+		when(restClient.getLeasePolicy("df60ff9e-4027-12d1-a2b5-5229b3cee282")).thenReturn(policy2);
+		when(restClient.getLeasePolicy("df60ff9e-4027-13d1-a2b5-5229b3cee282")).thenReturn(policy3);
+		when(restClient.getLeasePolicy("df60ff9e-4027-14d1-a2b5-5229b3cee282")).thenReturn(policy4);
+		when(restClient.getLeasePolicy("df60ff9e-4027-15d1-a2b5-5229b3cee282")).thenReturn(policy5);
 
 		File policyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, resourceQuotaPolicy).toFile();
+				.get(tempFolder.getRoot().getPath(), dirPolicies, leasePolicy).toFile();
 
 		// TEST
 		store.exportContent();
 
 		// VERIFY
 		assertEquals(6, Objects.requireNonNull(policyFolder.listFiles()).length);
-		AssertionsHelper.assertFolderContainsFiles(policyFolder, new String[] { "RQ01.json", "RQ01_1.json",
-				"RQ01_2.json", "RQ01_3.json", "RQ01_4.json", "RQ01_5.json" });
+		AssertionsHelper.assertFolderContainsFiles(policyFolder, new String[] { "LP01.json", "LP01_1.json",
+				"LP01_2.json", "LP01_3.json", "LP01_4.json", "LP01_5.json" });
 	}
 }
