+++ common/artifact-manager/src/test/java/com/vmware/pscoe/iac/artifact/aria/automation/store/VraNgResourceQuotaPolicyStoreTest.java
@@ -21,6 +21,7 @@ import com.vmware.pscoe.iac.artifact.helpers.FsMocks;
 import com.vmware.pscoe.iac.artifact.model.Package;
 import com.vmware.pscoe.iac.artifact.model.PackageFactory;
 import com.vmware.pscoe.iac.artifact.model.PackageType;
+import com.vmware.pscoe.iac.artifact.aria.automation.store.models.*;
 import com.vmware.pscoe.iac.artifact.aria.automation.models.*;
 import com.vmware.pscoe.iac.artifact.aria.automation.rest.RestClientVraNg;
 import org.junit.Rule;
@@ -33,18 +34,15 @@ import org.mockito.Mockito;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Paths;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.Objects;
+import java.util.*;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
-import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.when;
 
-public class VraNgDay2ActionsPolicyStoreTest {
+public class VraNgResourceQuotaPolicyStoreTest {
+
 	/**
 	 * Temp Folder.
 	 */
@@ -53,7 +51,7 @@ public class VraNgDay2ActionsPolicyStoreTest {
 	/**
 	 * store.
 	 */
-	private VraNgDay2ActionsPolicyStore store;
+	private VraNgResourceQuotaPolicyStore store;
 	/**
 	 * restClient.
 	 */
@@ -79,9 +77,9 @@ public class VraNgDay2ActionsPolicyStoreTest {
 	 */
 	private String dirPolicies = "policies";
 	/**
-	 * day2ActionsPolicy.
+	 * resourceQuotaPolicy.
 	 */
-	private String day2ActionsPolicy = "day2-actions";
+	private String resourceQuotaPolicy = "resource-quota";
 
 	/**
 	 * Init function called before each test.
@@ -95,7 +93,7 @@ public class VraNgDay2ActionsPolicyStoreTest {
 		}
 
 		fsMocks = new FsMocks(tempFolder.getRoot());
-		store = new VraNgDay2ActionsPolicyStore();
+		store = new VraNgResourceQuotaPolicyStore();
 		restClient = Mockito.mock(RestClientVraNg.class);
 		pkg = PackageFactory.getInstance(PackageType.VRANG, tempFolder.getRoot());
 		config = Mockito.mock(ConfigurationVraNg.class);
@@ -127,180 +125,174 @@ public class VraNgDay2ActionsPolicyStoreTest {
 	}
 
 	@Test
-	void testExportContentWithNoDay2ActionsPolicies() {
-		System.out.println(this.getClass() + "testExportContentWithNoDay2ActionsPolicies");
+	void testExportContentWithNoResourceQuotaPolicies() {
+		System.out.println(this.getClass() + "testExportContentWithNoResourceQuotaPolicies");
 		// GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(new VraNgPolicy());
 
 		// TEST
 		store.exportContent();
 
-		File day2ActionsPolicyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, day2ActionsPolicy).toFile();
+		File resourceQuotaPolicyFolder = Paths
+				.get(tempFolder.getRoot().getPath(), dirPolicies, resourceQuotaPolicy).toFile();
 
 		// VERIFY
-		verify(restClient, never()).getDay2ActionsPolicies();
-		verify(restClient, never()).getDay2ActionsPolicy(anyString());
+		verify(restClient, never()).getResourceQuotaPolicies();
+		verify(restClient, never()).getResourceQuotaPolicy(anyString());
 
-		assertEquals(null, day2ActionsPolicyFolder.listFiles());
+		assertEquals(null, resourceQuotaPolicyFolder.listFiles());
 	}
 
 	@Test
-	void testExportContentWithAllDay2ActionsPolicies() {
-		System.out.println(this.getClass() + "testExportContentWithAllDay2ActionsPolicies");
-		VraNgDay2ActionsPolicy policy1 = new VraNgDay2ActionsPolicy(
+	void testExportContentWithAllResourceQuotaPolicies() {
+		System.out.println(this.getClass() + "testExportContentWithAllResourceQuotaPolicies");
+		VraNgResourceQuotaPolicy rqPolicy = new VraNgResourceQuotaPolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"D2A01",
-				"com.vmware.policy.deployment.action",
+				"RQ01",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"b2c558c8-f20c-4da6-9bc3-d7561f64df16",
 				"HARD",
 				"TEST",
 				new JsonObject(),
-				new JsonObject(),
 				new JsonObject());
 
-		VraNgDay2ActionsPolicy policy2 = new VraNgDay2ActionsPolicy(
+		VraNgResourceQuotaPolicy rqPolicy2 = new VraNgResourceQuotaPolicy(
 				"2cf93725-38e9-4cb9-888a-a40994754c31",
-				"D2A02",
-				"com.vmware.policy.deployment.action",
+				"RQ02",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"b2c558c8-f20c-4da6-9bc3-d7561f64df16",
 				"HARD",
 				"TEST",
 				new JsonObject(),
-				new JsonObject(),
 				new JsonObject());
 
-		List<VraNgDay2ActionsPolicy> policies = Arrays.asList(policy1, policy2);
+		List<VraNgResourceQuotaPolicy> policies = Arrays.asList(rqPolicy, rqPolicy2);
 
 		// // GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(new VraNgPolicy(null, null, null, null, null, null));
-		when(restClient.getDay2ActionsPolicies()).thenReturn(policies);
-		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy1);
-		when(restClient.getDay2ActionsPolicy("2cf93725-38e9-4cb9-888a-a40994754c31")).thenReturn(policy2);
+		when(restClient.getResourceQuotaPolicies()).thenReturn(policies);
+		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(rqPolicy);
+		when(restClient.getResourceQuotaPolicy("2cf93725-38e9-4cb9-888a-a40994754c31")).thenReturn(rqPolicy2);
 
 		// TEST
 		store.exportContent();
 
 		// VERIFY
-		File day2ActionsPolicyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, day2ActionsPolicy).toFile();
-		assertEquals(2, Objects.requireNonNull(day2ActionsPolicyFolder.listFiles()).length);
+		File resourceQuotaPolicyFolder = Paths
+				.get(tempFolder.getRoot().getPath(), dirPolicies, resourceQuotaPolicy).toFile();
+		assertEquals(2, Objects.requireNonNull(resourceQuotaPolicyFolder.listFiles()).length);
 	}
 
 	@Test
-	void testExportContentWithSpecificDay2ActionsPolicies() {
-		System.out.println(this.getClass() + "testExportContentWithSpecificDay2ActionsPolicies");
+	void testExportContentWithSpecificResourceQuotaPolicies() {
+		System.out.println(this.getClass() + "testExportContentWithSpecificResourceQuotaPolicies");
 
-		VraNgDay2ActionsPolicy policy = new VraNgDay2ActionsPolicy(
+		VraNgResourceQuotaPolicy rqPolicy = new VraNgResourceQuotaPolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"D2A01",
-				"com.vmware.policy.deployment.action",
+				"RQ01",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST",
 				new JsonObject(),
-				new JsonObject(),
 				new JsonObject());
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, Collections.singletonList("D2A01"), null, null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, Collections.singletonList("RQ01"), null, null, null, null);
 		// // GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
-		when(restClient.getDay2ActionsPolicies()).thenReturn(Collections.singletonList(policy));
-		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
+		when(restClient.getResourceQuotaPolicies()).thenReturn(Collections.singletonList(rqPolicy));
+		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(rqPolicy);
 
 		// TEST
 		store.exportContent();
 
-		File day2ActionsPolicyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, day2ActionsPolicy).toFile();
+		File resourceQuotaPolicyFolder = Paths
+				.get(tempFolder.getRoot().getPath(), dirPolicies, resourceQuotaPolicy).toFile();
 
 		// VERIFY
-		assertEquals(1, Objects.requireNonNull(day2ActionsPolicyFolder.listFiles()).length);
+		assertEquals(1, Objects.requireNonNull(resourceQuotaPolicyFolder.listFiles()).length);
 	}
 
 	@Test
 	void testImportContentWithUpdateLogic() {
 		System.out.println("testImportContentWithUpdateLogic");
-		VraNgDay2ActionsPolicy policy = new VraNgDay2ActionsPolicy(
+		VraNgResourceQuotaPolicy rqPolicy = new VraNgResourceQuotaPolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"D2A01",
-				"com.vmware.policy.deployment.action",
+				"RQ01",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST",
 				new JsonObject(),
-				new JsonObject(),
 				new JsonObject());
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, Collections.singletonList("D2A01"), null, null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, Collections.singletonList("RQ01"), null, null, null, null);
 		// GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
 
-		fsMocks.getDay2ActionsPolicyFsMocks().addPolicy(policy);
+		fsMocks.resourceQuotaPolicyFsMocks().addResourceQuotaPolicy(rqPolicy);
 
-		File day2ActionsPolicyFolder = Paths
-				.get(fsMocks.getTempFolderProjectPath().getPath(), day2ActionsPolicy).toFile();
+		File resourceQuotaPolicyFolder = Paths
+				.get(fsMocks.getTempFolderProjectPath().getPath(), resourceQuotaPolicy).toFile();
 
-		AssertionsHelper.assertFolderContainsFiles(day2ActionsPolicyFolder, new String[] { "D2A01.json" });
+		AssertionsHelper.assertFolderContainsFiles(resourceQuotaPolicyFolder, new String[] { "RQ01.json" });
 
-		when(restClient.getDay2ActionsPolicies()).thenReturn(Arrays.asList(policy));
-		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
+		when(restClient.getResourceQuotaPolicies()).thenReturn(Arrays.asList(rqPolicy));
+		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(rqPolicy);
 
 		// START TEST
 		store.importContent(tempFolder.getRoot());
 
 		// VERIFY
-		verify(restClient, times(1)).createDay2ActionsPolicy(any());
+		verify(restClient, times(1)).createResourceQuotaPolicy(any());
 	}
 
 	@Test
 	void testImportContentWithCreateLogic() {
 		System.out.println("testImportContentWithCreateLogic");
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, Arrays.asList("D2A01"), null, null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, Collections.singletonList("RQ01"), null, null, null, null);
 
-		VraNgDay2ActionsPolicy policy = new VraNgDay2ActionsPolicy(
+		VraNgResourceQuotaPolicy rqPolicy = new VraNgResourceQuotaPolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"D2A01",
-				"com.vmware.policy.deployment.action",
+				"RQ01",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"b2c558c8-f20c-4da6-9bc3-d7561f64df16",
 				"HARD",
 				"TEST",
-				new JsonObject(),
 				new JsonObject(),
 				new JsonObject());
-		VraNgDay2ActionsPolicy policyFromServer = new VraNgDay2ActionsPolicy(
+		VraNgResourceQuotaPolicy rqPolicyFromServer = new VraNgResourceQuotaPolicy(
 				"2cf93725-38e9-4cb9-888a-a40994754c31",
-				"D2A02",
-				"com.vmware.policy.deployment.action",
+				"RQ02",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"b2c558c8-f20c-4da6-9bc3-d7561f64df16",
 				"HARD",
 				"TEST",
-				new JsonObject(),
 				new JsonObject(),
 				new JsonObject());
 
 		// GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
 
-		fsMocks.getDay2ActionsPolicyFsMocks().addPolicy(policy);
+		fsMocks.resourceQuotaPolicyFsMocks().addResourceQuotaPolicy(rqPolicy);
 
-		File day2ActionsPolicyFolder = Paths
-				.get(fsMocks.getTempFolderProjectPath().getPath(), day2ActionsPolicy).toFile();
+		File resourceQuotaPolicyFolder = Paths
+				.get(fsMocks.getTempFolderProjectPath().getPath(), resourceQuotaPolicy).toFile();
 
-		AssertionsHelper.assertFolderContainsFiles(day2ActionsPolicyFolder, new String[] { "D2A01.json" });
+		AssertionsHelper.assertFolderContainsFiles(resourceQuotaPolicyFolder, new String[] { "RQ01.json" });
 
-		when(restClient.getDay2ActionsPolicies()).thenReturn(Collections.emptyList());
-		when(restClient.getDay2ActionsPolicy("2cf93725-38e9-4cb9-888a-a40994754c31")).thenReturn(policyFromServer);
+		when(restClient.getResourceQuotaPolicies()).thenReturn(Collections.emptyList());
+		when(restClient.getResourceQuotaPolicy("2cf93725-38e9-4cb9-888a-a40994754c31")).thenReturn(rqPolicyFromServer);
 
 		// START TEST
 		store.importContent(tempFolder.getRoot());
 
 		// VERIFY
-		verify(restClient, times(1)).createDay2ActionsPolicy(any());
+		verify(restClient, times(1)).createResourceQuotaPolicy(any());
 	}
 
 	@Test
@@ -312,36 +304,35 @@ public class VraNgDay2ActionsPolicyStoreTest {
 
 		// VERIFY
 		verify(vraNgPackageDescriptor, never()).getPolicy();
-		verify(restClient, never()).getDay2ActionsPolicies();
-		verify(restClient, never()).getDay2ActionsPolicy(anyString());
-		verify(restClient, never()).createDay2ActionsPolicy(any());
+		verify(restClient, never()).getResourceQuotaPolicies();
+		verify(restClient, never()).getResourceQuotaPolicy(anyString());
+		verify(restClient, never()).createResourceQuotaPolicy(any());
 	}
 
 	@Test
 	void testExportContentWithPolicyAlreadyInFile() {
 		System.out.println(this.getClass() + ".testExportContentWithPolicyAlreadyInFile");
 
-		VraNgDay2ActionsPolicy policy = new VraNgDay2ActionsPolicy(
+		VraNgResourceQuotaPolicy policy = new VraNgResourceQuotaPolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"D2A01",
-				"com.vmware.policy.deployment.action",
+				"RQ01",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST",
 				new JsonObject(),
-				new JsonObject(),
 				new JsonObject());
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, Collections.singletonList("D2A01"), null, null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, Collections.singletonList("RQ01"), null, null, null, null);
 		// // GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
-		when(restClient.getDay2ActionsPolicies()).thenReturn(Collections.singletonList(policy));
-		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
+		when(restClient.getResourceQuotaPolicies()).thenReturn(Collections.singletonList(policy));
+		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
 
 		File policyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, day2ActionsPolicy).toFile();
+				.get(tempFolder.getRoot().getPath(), dirPolicies, resourceQuotaPolicy).toFile();
 
-		fsMocks.getDay2ActionsPolicyFsMocks().addPolicy(policy);
+		fsMocks.resourceQuotaPolicyFsMocks().addResourceQuotaPolicy(policy);
 		// TEST
 		store.exportContent();
 
@@ -353,45 +344,43 @@ public class VraNgDay2ActionsPolicyStoreTest {
 	@Test
 	void testExportContentWithSpecificPoliciesAndDuplicateFiles() {
 		System.out.println(this.getClass() + ".testExportContentWithSpecificPoliciesAndDuplicateFiles");
-		VraNgDay2ActionsPolicy policyInFile = new VraNgDay2ActionsPolicy(
+		VraNgResourceQuotaPolicy policyInFile = new VraNgResourceQuotaPolicy(
 				"d160119e-4027-48d1-a2b5-5229b3cee282",
-				"D2A01",
-				"com.vmware.policy.deployment.action",
+				"RQ01",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"SOFT",
 				"TEST",
 				new JsonObject(),
-				new JsonObject(),
 				new JsonObject());
 
-		VraNgDay2ActionsPolicy policy = new VraNgDay2ActionsPolicy(
+		VraNgResourceQuotaPolicy policy = new VraNgResourceQuotaPolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"D2A01",
-				"com.vmware.policy.deployment.action",
+				"RQ01",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST",
 				new JsonObject(),
-				new JsonObject(),
 				new JsonObject());
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, Collections.singletonList("D2A01"), null, null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, Collections.singletonList("RQ01"), null, null, null, null);
 		// // GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
-		when(restClient.getDay2ActionsPolicies()).thenReturn(Collections.singletonList(policy));
-		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
+		when(restClient.getResourceQuotaPolicies()).thenReturn(Collections.singletonList(policy));
+		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
 
 		File policyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, day2ActionsPolicy).toFile();
+				.get(tempFolder.getRoot().getPath(), dirPolicies, resourceQuotaPolicy).toFile();
 
-		fsMocks.getDay2ActionsPolicyFsMocks().addPolicy(policyInFile);
-		policyInFile.setName("D2A01_1");
-		fsMocks.getDay2ActionsPolicyFsMocks().addPolicy(policyInFile);
-		policyInFile.setName("D2A01_2");
-		fsMocks.getDay2ActionsPolicyFsMocks().addPolicy(policyInFile);
-		policyInFile.setName("D2A01_3");
-		fsMocks.getDay2ActionsPolicyFsMocks().addPolicy(policyInFile);
+		fsMocks.resourceQuotaPolicyFsMocks().addResourceQuotaPolicy(policyInFile);
+		policyInFile.setName("RQ01_1");
+		fsMocks.resourceQuotaPolicyFsMocks().addResourceQuotaPolicy(policyInFile);
+		policyInFile.setName("RQ01_2");
+		fsMocks.resourceQuotaPolicyFsMocks().addResourceQuotaPolicy(policyInFile);
+		policyInFile.setName("RQ01_3");
+		fsMocks.resourceQuotaPolicyFsMocks().addResourceQuotaPolicy(policyInFile);
 
 		// TEST
 		store.exportContent();
@@ -399,113 +388,105 @@ public class VraNgDay2ActionsPolicyStoreTest {
 		// VERIFY
 		assertEquals(5, Objects.requireNonNull(policyFolder.listFiles()).length);
 		AssertionsHelper.assertFolderContainsFiles(policyFolder,
-				new String[] { "D2A01.json", "D2A01_1.json", "D2A01_2.json", "D2A01_3.json", "D2A01_4.json" });
+				new String[] { "RQ01.json", "RQ01_1.json", "RQ01_2.json", "RQ01_3.json", "RQ01_4.json" });
 	}
 
 	@Test
 	void testExportContentWithSpecificPoliciesAndDuplicateNames() {
 		System.out.println(this.getClass() + ".testExportContentWithSpecificPoliciesAndDuplicateNames");
-		VraNgDay2ActionsPolicy policyInFile = new VraNgDay2ActionsPolicy(
+		VraNgResourceQuotaPolicy policyInFile = new VraNgResourceQuotaPolicy(
 				"d160119e-4027-48d1-a2b5-5229b3cee282",
-				"D2A01",
-				"com.vmware.policy.deployment.action",
+				"RQ01",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"SOFT",
 				"TEST",
 				new JsonObject(),
-				new JsonObject(),
 				new JsonObject());
 
-		VraNgDay2ActionsPolicy policy = new VraNgDay2ActionsPolicy(
+		VraNgResourceQuotaPolicy policy = new VraNgResourceQuotaPolicy(
 				"df60ff9e-4027-48d1-a2b5-5229b3cee282",
-				"D2A01",
-				"com.vmware.policy.deployment.action",
+				"RQ01",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST",
 				new JsonObject(),
-				new JsonObject(),
 				new JsonObject());
-		VraNgDay2ActionsPolicy policy1 = new VraNgDay2ActionsPolicy(
+		VraNgResourceQuotaPolicy policy1 = new VraNgResourceQuotaPolicy(
 				"df60ff9e-4027-11d1-a2b5-5229b3cee282",
-				"D2A01",
-				"com.vmware.policy.deployment.action",
+				"RQ01",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST1",
 				new JsonObject(),
-				new JsonObject(),
 				new JsonObject());
-		VraNgDay2ActionsPolicy policy2 = new VraNgDay2ActionsPolicy(
+		VraNgResourceQuotaPolicy policy2 = new VraNgResourceQuotaPolicy(
 				"df60ff9e-4027-12d1-a2b5-5229b3cee282",
-				"D2A01",
-				"com.vmware.policy.deployment.action",
+				"RQ01",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST2",
 				new JsonObject(),
-				new JsonObject(),
 				new JsonObject());
-		VraNgDay2ActionsPolicy policy3 = new VraNgDay2ActionsPolicy(
+		VraNgResourceQuotaPolicy policy3 = new VraNgResourceQuotaPolicy(
 				"df60ff9e-4027-13d1-a2b5-5229b3cee282",
-				"D2A01",
-				"com.vmware.policy.deployment.action",
+				"RQ01",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST3",
 				new JsonObject(),
-				new JsonObject(),
 				new JsonObject());
-		VraNgDay2ActionsPolicy policy4 = new VraNgDay2ActionsPolicy(
+		VraNgResourceQuotaPolicy policy4 = new VraNgResourceQuotaPolicy(
 				"df60ff9e-4027-14d1-a2b5-5229b3cee282",
-				"D2A01",
-				"com.vmware.policy.deployment.action",
+				"RQ01",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST4",
 				new JsonObject(),
-				new JsonObject(),
 				new JsonObject());
-		VraNgDay2ActionsPolicy policy5 = new VraNgDay2ActionsPolicy(
+		VraNgResourceQuotaPolicy policy5 = new VraNgResourceQuotaPolicy(
 				"df60ff9e-4027-15d1-a2b5-5229b3cee282",
-				"D2A01",
-				"com.vmware.policy.deployment.action",
+				"RQ01",
+				"com.vmware.policy.resource.quota",
 				"b899c648-bf84-4d35-a61c-db212ecb4c1e",
 				"VIDM-L-01A",
 				"HARD",
 				"TEST5",
 				new JsonObject(),
-				new JsonObject(),
 				new JsonObject());
 
-		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, null, Arrays.asList("D2A01"), null, null, null);
+		VraNgPolicy vraNgPolicy = new VraNgPolicy(null, Collections.singletonList("RQ01"), null, null, null, null);
 		// // GIVEN
 		when(vraNgPackageDescriptor.getPolicy()).thenReturn(vraNgPolicy);
-		when(restClient.getDay2ActionsPolicies())
+		when(restClient.getResourceQuotaPolicies())
 				.thenReturn(Arrays.asList(policy, policy1, policy2, policy3, policy4, policy5));
-		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
-		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-11d1-a2b5-5229b3cee282")).thenReturn(policy1);
-		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-12d1-a2b5-5229b3cee282")).thenReturn(policy2);
-		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-13d1-a2b5-5229b3cee282")).thenReturn(policy3);
-		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-14d1-a2b5-5229b3cee282")).thenReturn(policy4);
-		when(restClient.getDay2ActionsPolicy("df60ff9e-4027-15d1-a2b5-5229b3cee282")).thenReturn(policy5);
+		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-48d1-a2b5-5229b3cee282")).thenReturn(policy);
+		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-11d1-a2b5-5229b3cee282")).thenReturn(policy1);
+		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-12d1-a2b5-5229b3cee282")).thenReturn(policy2);
+		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-13d1-a2b5-5229b3cee282")).thenReturn(policy3);
+		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-14d1-a2b5-5229b3cee282")).thenReturn(policy4);
+		when(restClient.getResourceQuotaPolicy("df60ff9e-4027-15d1-a2b5-5229b3cee282")).thenReturn(policy5);
 
 		File policyFolder = Paths
-				.get(tempFolder.getRoot().getPath(), dirPolicies, day2ActionsPolicy).toFile();
+				.get(tempFolder.getRoot().getPath(), dirPolicies, resourceQuotaPolicy).toFile();
 
 		// TEST
 		store.exportContent();
 
 		// VERIFY
 		assertEquals(6, Objects.requireNonNull(policyFolder.listFiles()).length);
-		AssertionsHelper.assertFolderContainsFiles(policyFolder, new String[] { "D2A01.json", "D2A01_1.json",
-				"D2A01_2.json", "D2A01_3.json", "D2A01_4.json", "D2A01_5.json" });
+		AssertionsHelper.assertFolderContainsFiles(policyFolder, new String[] { "RQ01.json", "RQ01_1.json",
+				"RQ01_2.json", "RQ01_3.json", "RQ01_4.json", "RQ01_5.json" });
 	}
-
 }
