+++ common/artifact-manager/src/test/java/com/vmware/pscoe/iac/artifact/aria/automation/utils/VraNgProjectUtilTest.java
